requirements.txt
streamlit
google-generativeai
google-cloud-vision
python-dotenv
Pillow
watchdog

